Moved Done menu to shipping fragment
